hotfix bugs

+++ src/conditions_01.c
@@ -35,9 +35,7 @@ int check_numbers_condtions(size_t line_num)
  * Return: void
  */
 
-
-/*
-int condition(size_t line_num)
+size_t condition(size_t line_num)
 {
 	int count_condtions = 0;
 	size_t new_line = 0;
@@ -51,4 +49,3 @@ int condition(size_t line_num)
 
 	return (new_line);
 }
-*/
